Consolidate agent diagnostics and restore deleted review coverage
Replace ViftyAgentRule, ViftyCtlReadinessReport, and ViftyCtlRecoverySteps
with a single AgentDiagnostics module (same public symbols, consolidated
readiness/recovery/rule logic), keep the full documentation/evidence test
suite in place, and update CODEOWNERS plus AGENTS.md to the consolidated
layout. The advisory update checker remains the documented Developer-ID
release-page handoff; no in-app download/install updater is introduced.
Revert the worktree's accidental deletion of DocumentationTrustSurfaceTests,
UIReviewEvidenceScriptTests, and the trimmed release-metadata/validation
coverage, restoring the Makefile slow-suite skip list and trust-gate
assertions that pin them.
